Add: redis_writer send data to redis

+++ src/util.h
@@ -12,7 +12,7 @@ namespace Util {
 // sock util
 sockaddr_in NewSockaddrInet(const std::string &host, uint32_t port);
 Status SockConnect(std::string host, uint32_t port, int *fd);
-int SockSend(int fd, const std::string &data);
+ssize_t SockSend(int fd, const std::string &data);
 int GetPeerAddr(int fd, std::string *addr, uint32_t *port);
 bool IsPortInUse(int port);
